text-formatting: margins: cleanup
- change valid-margin-spec-p

add a destructuring-bind to enforce two elements on the list and we
check if the second value is a real number (not only non-nil).

- add normalize-margin-spec utility

function replaces two local macrolets. It takes the margin
specification plist and a plist with default values for said plist and
normalizes it as follows:

1. when edge is not specified it is taken from defaults
2. when edge value is an atom replace it with (:relative atom)
3. otherwise do nothing - edge value is assumed to be correct

Finally it checks if the resulting spec is valid. Function does not
modify its arguments.